+++ package.json
@@ -28,9 +28,9 @@
     "@upstash/ratelimit": "^0.4.4",
     "@upstash/redis": "^1.22.0",
     "autoprefixer": "^10.4.15",
+    "csv42": "^4.0.0",
     "dotenv": "^16.3.1",
     "drizzle-orm": "^0.28.5",
-    "json2csv": "6.0.0-alpha.2",
     "next": "^13.4.19",
     "nextjs-cors": "^2.1.2",
     "postcss": "^8.4.29",
