+++ src/app/dashboard/page.tsx
@@ -4,9 +4,8 @@ import { GiStarFormation } from "react-icons/gi";
 import { isAdmin, isVIP } from "@/utils/helpers";
 import { currentUser } from "@clerk/nextjs";
 
-export const dynamic = "force-dynamic";
-export const revalidate = 0;
-export const fetchCache = "force-no-store";
+export const runtime = "edge";
+export const preferredRegion = ["pdx1"];
 
 export default async function Dashboard() {
   const user = await currentUser();
